requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.23.0
google-cloud-bigquery[pandas]==1.20.0
Flask==1.1.1
    

